SystemUI: change default fingerprint-behavior
Backported shouldListenForFingerprint() from cm-14.1

This fixes unexpected fingerprint-authentication attempts even when the screen is off.

Additionally, allow device to specifically allow fingerprint to authenticate when
either going to sleep or when draming.

A possible side-effects this commit may have on improper codec fingerprint-HALs is that if it doesn't
respond with FINGERPRINT_ERROR_CANCELED to an cancel()-request, FingerprintService doesn't properly
handle the cancel and thus cannot start new attempts afterwards:

      03-07 11:58:07.158  3485  3485 W FingerprintService: client com.android.systemui is no longer authenticating
      //
      // no working fingerprint-sensor in here, even when screen is turned on/off several times
      //
      03-07 11:58:33.877  3485  3485 V FingerprintService: startAuthentication(com.android.systemui)
      03-07 11:58:33.877  3485  3485 V FingerprintService: request stop current client com.android.systemui
      03-07 11:58:33.877  3485  3485 W FingerprintService: stopAuthentication: already cancelled!
      03-07 11:58:36.879  3485  3485 W FingerprintService: Client com.android.systemui failed to respond to cancel, starting client com.android.systemui
      03-07 11:58:36.879  3485  3485 V FingerprintService: starting client AuthenticationClient(com.android.systemui), initiatedByClient = false)

This bug was only observed on zero (Samsung Galaxy S6) yet, a fix for this behavior has already been submitted
to the hardware/samsung/-repo. This bugs "abuses" the already exisiting notify-wrapper to send a custom
error-message with the code FINGERPRINT_ERROR_CANCELED to Android.

Change-Id: I94eac4b1586cdae4454d7a233d87626664f667a6